Update dark mode switcher and wezterm install

+++ wezterm/install.sh
@@ -1,1 +1,5 @@
-ln -s ~/.dotfiles/wezterm/.wezterm.lua ~/.wezterm.lua
+#!/bin/bash
+
+if [ ! -f ~/.wezterm.lua ]; then
+	ln -s ~/.dotfiles/wezterm/.wezterm.lua ~/.wezterm.lua
+fi
